Follow Apache XMLRPC2->XMLRPC3 API changes
The 'SAMPManager' components of Aladin were written
against the Apache XMLRPC v2 API; but only the v3 libraries
are available in Debian.  The APIs been altered both in
namespace, and also in operation.  Unfortunately
nearly all documentation and examples are only for v2;
so this patch has been developed somewhat blindly.