Patch: CDS Astro Proj3.set() API
The source for CDS Astro is missing, and the retrieved source code rescued
from earlier versions appears to have a slightly different API.  However in
each of the conflicting cases, a commented out alternative was found in the
Aladin source that appears to work when uncommented and used instead.
It is unclear which is the 'newer'/prefered API variant.